NavigationView used with menu instead of complicted views and stuff
