+++ PosseUpApp/app/build.gradle
@@ -30,12 +30,12 @@ dependencies {
     compile 'com.rengwuxian.materialedittext:library:2.1.4'
     compile 'com.mikhaellopez:circularimageview:2.1.1'
     compile 'com.microsoft.azure:azure-mobile-services-android-sdk:2.0.3'
-    compile 'com.google.android.gms:play-services-location:8.1.0'
-    compile 'com.google.android.gms:play-services-nearby:8.1.0'
-    compile 'com.google.android.gms:play-services-gcm:8.1.0'
+    compile 'com.google.android.gms:play-services-location:8.4.0'
+    compile 'com.google.android.gms:play-services-nearby:8.4.0'
+    compile 'com.google.android.gms:play-services-gcm:8.4.0'
     compile 'com.android.support:support-v4:23.1.1'
     compile files('libs/notification-hubs-0.4.jar')
-    compile 'com.google.android.gms:play-services-base:6.5.87'
+    compile 'com.google.android.gms:play-services-base:8.4.0'
     compile 'com.microsoft.azure:azure-notifications-handler:1.0.1@aar'
 }
 repositories{
