+++ PosseUpApp/app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:theme="@style/AppTheme" >
 
         <activity
-            android:name=".UI.MainActivity"
+            android:name=".UI.MainActivityGroup.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/RevisedAppTheme">
