+++ PosseUpApp/app/src/main/java/com/lordan/mark/PosseUp/AbstractActivity.java
@@ -19,7 +19,7 @@ public abstract class AbstractActivity extends ActionBarActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setRequestedOrientation (ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
+//        setRequestedOrientation (ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
     }
     protected boolean isValidEmail(String target){
         if (TextUtils.isEmpty(target)) {
@@ -59,6 +59,13 @@ public abstract class AbstractActivity extends ActionBarActivity {
         SharedPreferences prefs = getApplicationContext().getSharedPreferences("PosseUpData", MODE_PRIVATE);
         if(prefs != null){
             return prefs.getString("email", null);
+        }
+        return null;
+    }
+    protected String getCurrentUsername(){
+        SharedPreferences prefs = getApplicationContext().getSharedPreferences("PosseUpData", MODE_PRIVATE);
+        if(prefs != null){
+            return prefs.getString("username", null);
         }
         return null;
     }
