+++ PosseUpApp/app/src/main/java/com/lordan/mark/PosseUp/UI/RegisterActivity.java
@@ -25,6 +25,7 @@ import com.lordan.mark.PosseUp.DataOperations.AzureService;
 import com.lordan.mark.PosseUp.Model.Constants;
 import com.lordan.mark.PosseUp.Model.User;
 import com.lordan.mark.PosseUp.R;
+import com.lordan.mark.PosseUp.UI.MainActivityGroup.MainActivity;
 
 import org.json.JSONArray;
 import org.json.JSONException;
