+++ PosseUpApp/app/src/main/java/com/lordan/mark/PosseUp/UI/SigninGroup/SigninFrag.java
@@ -5,11 +5,8 @@ import android.app.Fragment;
 import android.app.ProgressDialog;
 
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
 
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -34,17 +31,15 @@ import com.lordan.mark.PosseUp.DataOperations.AzureService;
 import com.lordan.mark.PosseUp.Model.Constants;
 
 import com.lordan.mark.PosseUp.R;
-import com.lordan.mark.PosseUp.UI.MainActivity;
+import com.lordan.mark.PosseUp.UI.MainActivityGroup.MainActivity;
 import com.lordan.mark.PosseUp.UI.RegisterActivity;
 
 
-import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.HashMap;
 import java.util.Map;
-import java.util.jar.Manifest;
 
 /**
  * Created by Mark on 10/10/2015.
