+++ PosseUpApp/app/src/main/res/color/drawer_selector.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <!-- This is used when the Navigation Item is checked -->
+    <item android:color="#000" android:state_checked="true" />
+    <!-- This is the default text color -->
+    <item android:color="#757575" />
+</selector>
