+++ PosseUpApp/app/src/main/res/layout/drawer_header_layout.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout
+    android:layout_width="match_parent"
+    android:layout_height="128dp"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:background="@drawable/gradient_background"
+    android:paddingBottom="@dimen/margin_small"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_marginStart="@dimen/margin_medium"
+        android:layout_marginTop="@dimen/margin_medium"
+        android:orientation="vertical">
+        <com.mikhaellopez.circularimageview.CircularImageView
+            android:layout_width="65dp"
+            android:layout_height="65dp"
+            app:border="false"
+            android:src="@drawable/profiler"/>
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="56dp"
+            android:orientation="vertical">
+            <TextView
+                android:id="@+id/drawer_username"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="Username"
+                android:fontFamily="sans-serif-medium"
+                android:textSize="@dimen/abc_text_size_small_material"
+                android:layout_marginTop="@dimen/margin_tiny"
+                android:textColor="@color/white"/>
+            <TextView
+                android:id="@+id/drawer_email"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="Email Address"
+                android:fontFamily="sans-serif-regular"
+                android:textSize="@dimen/abc_text_size_small_material"/>
+        </LinearLayout>
+
+    </LinearLayout>
+</FrameLayout>
