+++ PosseUpApp/app/src/main/res/values/color.xml
@@ -7,6 +7,6 @@
     <color name="white">#FFF8F8</color>
     <color name="primaryThemeColor">#FF3366</color>
 
-    <color name="grey">#666</color>
+    <color name="grey">#757575</color>
 
 </resources>
