+++ PosseUpApp/app/src/main/res/values/strings.xml
@@ -41,6 +41,14 @@
     <string name="closeddrawer">Closed</string>
 
     <!-- Nav Drawer Menu Items -->
+    <string name="drawer_home">Home</string>
+    <string name="drawer_profile">Profile</string>
+    <string name="drawer_events">Events</string>
+    <string name="drawer_friends">Friends</string>
+    <string name="drawer_settings">Settings</string>
+
+
+
     <string-array name="nav_drawer_items">
         <item >Home</item>
         <item >Profile</item>
