+++ PosseUpApp/app/src/main/res/values/styles.xml
@@ -7,6 +7,7 @@
     </style>
 
     <style name="RevisedAppTheme" parent="@style/Theme.AppCompat.NoActionBar">
+        <item name="android:listDivider">@color/grey</item>
     </style>
 
 </resources>
