+++ PosseUpApp/local.properties
@@ -7,5 +7,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Sat Jan 30 19:11:41 GMT 2016
-sdk.dir=C\:\\Users\\Mark\\AppData\\Local\\Android\\sdk
+#Mon Feb 01 09:08:58 GMT 2016
+sdk.dir=/Users/Mark/Library/Android/sdk
